# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Booking App",
        "description": "This is the Event Booking App - API Documentation generated using Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "EventApp API Server"
        }
    ],
    "paths": {
        "/api/book-event": {
            "post": {
                "tags": [
                    "Attendees APIs"
                ],
                "summary": "Book an event for an attendee",
                "description": "This API allows an attendee to book an event by providing details such as name, email, and event ID.",
                "operationId": "bookEvent",
                "requestBody": {
                    "description": "Booking details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "event_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name of the attendee.",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "The email address of the attendee.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "event_id": {
                                        "description": "The ID of the event being booked.",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "address": {
                                        "description": "The address of the attendee (optional).",
                                        "type": "string",
                                        "example": "1234 Elm Street"
                                    },
                                    "phone": {
                                        "description": "The phone number of the attendee (optional).",
                                        "type": "string",
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event successfully booked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Event booked successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "user_email": {
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                },
                                                "event": {
                                                    "type": "string",
                                                    "example": "Tech Conference"
                                                },
                                                "booking_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-04-22"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error or event already booked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Event ABC already booked by john.doe@example.com on 2025-04-22."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - required fields missing or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name field is required."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email field is required."
                                                    }
                                                },
                                                "event_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The event_id field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/get-my-booked-events": {
            "post": {
                "tags": [
                    "Attendees APIs"
                ],
                "summary": "Get a user's booked events",
                "description": "This API allows a user to get their booked events by providing their email and optionally a page number for pagination.",
                "operationId": "getMyBookedEvents",
                "requestBody": {
                    "description": "The user's email and page number for pagination",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "The email address of the user.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "page_number": {
                                        "description": "The page number for pagination (optional).",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booked events fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Booked events fetched successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "events": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "event_name": {
                                                                "type": "string",
                                                                "example": "Tech Conference 2025"
                                                            },
                                                            "event_date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2025-12-01"
                                                            },
                                                            "booking_date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2025-04-22"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "total_pages": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No events found for the user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No events found for user with email: john.doe@example.com."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - invalid email or page number",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email field is required."
                                                    }
                                                },
                                                "page_number": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The page number must be at least 1."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}